Fix Linux build break; simplify logic and avoid double-parsing in IntSliderViewItem::metadataChanged()

+++ ValueEditor/IntSliderViewItem.h
@@ -46,10 +46,10 @@ private:
   QWidget *m_widget;
   VELineEdit *m_lineEdit;
   QSlider *m_slider;
-  double m_softMinimum;
-  double m_softMaximum;
-  double m_hardMinimum;
-  double m_hardMaximum;
+  int m_softMinimum;
+  int m_softMaximum;
+  int m_hardMinimum;
+  int m_hardMaximum;
   bool m_isSettingValue;
 
 private slots:
